+++ CHANGELOG.md
@@ -7,9 +7,10 @@
 ### Added
 ### Changed
 
-* API: `font_` prefix changed to `fcft_`
-* API: renamed `struct glyph` to `struct fcft_glyph`
-* API: `enum subpixel_order` renamed to `enum fcft_subpixel`, and
+* API: `font_` prefix changed to `fcft_`.
+* API: renamed `struct font` to `struct fcft_font`.
+* API: renamed `struct glyph` to `struct fcft_glyph`.
+* API: renamed `enum subpixel_order` to `enum fcft_subpixel`, and
   `ORDER` was removed from the enum values.